requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
openai==1.63.0
httpx==0.25.1
python-dotenv==1.0.1
pytz==2024.1
SQLAlchemy==2.0.35
python-dateutil==2.8.2 